store: persist canonical raw event wire
- derive store ingest from verified signed event wire

- persist outbox signed and raw JSON from canonical NIP-01 material

- align downstream Nostr, trade, transport, and authority callers

- verify event-store/outbox tests plus contract and hygiene gates

+++ crates/authority/src/error.rs
@@ -59,8 +59,8 @@ pub enum RadrootsAuthorityError {
     },
 
     SignedEventCreatedAtMismatch {
-        expected_created_at: u32,
-        actual_created_at: u32,
+        expected_created_at: u64,
+        actual_created_at: u64,
     },
 
     SignedEventKindMismatch {
