+++ crates/nostr/src/job_adapter.rs
@@ -60,8 +60,8 @@ impl JobEventLike for RadrootsNostrEventAdapter<'_> {
     fn raw_author(&self) -> String {
         self.author_hex.clone()
     }
-    fn raw_published_at(&self) -> u32 {
-        self.evt.created_at.as_secs() as u32
+    fn raw_published_at(&self) -> u64 {
+        self.evt.created_at.as_secs()
     }
     fn raw_kind(&self) -> u32 {
         match self.evt.kind {
